Addons for the mclust package

Extend the functionality of the mclust package for Gaussian finite mixture modeling by including:

density estimation for data with bounded support (Scrucca, 2019)

modal clustering using MEM algorithm for Gaussian mixtures (Scrucca, 2021)

entropy estimation via Gaussian mixture modeling (Robin & Scrucca, 2023)

Details

For a quick introduction to mclustAddons see the vignette A quick tour of mclustAddons.

See also:

densityMclustBounded() for density estimation of bounded data;

MclustMEM() for modal clustering;

EntropyGMM() for entropy estimation.

Author

Luca Scrucca.

Maintainer: Luca Scrucca luca.scrucca@unipg.it

References

Scrucca L. (2019) A transformation-based approach to Gaussian mixture density estimation for bounded data. Biometrical Journal, 61:4, 873–888. https://doi.org/10.1002/bimj.201800174

Scrucca L. (2021) A fast and efficient Modal EM algorithm for Gaussian mixtures. Statistical Analysis and Data Mining, 14:4, 305–314. https://doi.org/10.1002/sam.11527

Robin S. and Scrucca L. (2023) Mixture-based estimation of entropy. Computational Statistics & Data Analysis, 177, 107582. https://doi.org/10.1016/j.csda.2022.107582